Move oEmbed endpoint under images
