+++ openverse-api/test/image_integration_test.py
@@ -4,9 +4,10 @@ functioning as designed. Run with the `pytest -s` command from this directory.
 """
 
 import json
-
+from urllib.parse import urlencode
 import pytest
 import requests
+import xml.etree.ElementTree as ET
 
 from test.constants import API_URL
 from test.media_integration import (
@@ -50,3 +51,40 @@ def test_image_detail(image_fixture):
 
 def test_image_stats():
     stats('images', 'image_count')
+
+
+def test_oembed_endpoint_for_json():
+    params = {
+        'url': 'https://any.domain/any/path/29cb352c-60c1-41d8-bfa1-7d6f7d955f63',
+        # 'format': 'json' is the default
+    }
+    response = requests.get(
+        f'{API_URL}/v1/images/oembed?{urlencode(params)}',
+        verify=False
+    )
+    assert response.status_code == 200
+    assert response.headers['Content-Type'] == "application/json"
+
+    parsed = response.json()
+    assert parsed['width'] == 1276
+    assert parsed['height'] == 1536
+    assert parsed['license_url'] == 'https://creativecommons.org/licenses/by-nc-nd/4.0/'
+
+
+def test_oembed_endpoint_for_xml():
+    params = {
+        'url': 'https://any.domain/any/path/29cb352c-60c1-41d8-bfa1-7d6f7d955f63',
+        'format': 'xml'
+    }
+    response = requests.get(
+        f'{API_URL}/v1/images/oembed?{urlencode(params)}',
+        verify=False
+    )
+    assert response.status_code == 200
+    assert response.headers['Content-Type'] == "application/xml; charset=utf-8"
+
+    response_body_as_xml = ET.fromstring(response.content)
+    xml_tree = ET.ElementTree(response_body_as_xml)
+    assert xml_tree.find("width").text == '1276'
+    assert xml_tree.find("height").text == '1536'
+    assert xml_tree.find("license_url").text == 'https://creativecommons.org/licenses/by-nc-nd/4.0/'
