+++ openverse-api/test/v1_integration_test.py
@@ -9,7 +9,6 @@ import pytest
 import uuid
 import time
 import catalog.settings
-import xml.etree.ElementTree as ET
 from django.db.models import Max
 from django.urls import reverse
 
@@ -456,36 +455,6 @@ def test_related_image_search_page_consistency(
         related = recommendation_factory(image['id'])
         assert related['result_count'] > 0
         assert len(related['results']) == 10
-
-
-def test_oembed_endpoint_for_json():
-    response = requests.get(
-        f'{API_URL}/v1/oembed?url=https%3A//'
-        'search.creativecommons.org/photos/'
-        '29cb352c-60c1-41d8-bfa1-7d6f7d955f63'
-    )
-    assert response.status_code == 200
-    assert response.headers['Content-Type'] == "application/json"
-    parsed = response.json()
-    assert parsed['width'] == 1276
-    assert parsed['height'] == 1536
-    assert parsed['license_url'] == 'https://creativecommons.org/licenses/by-nc-nd/4.0/'
-
-
-def test_oembed_endpoint_for_xml():
-    response = requests.get(
-        f'{API_URL}/v1/oembed?url=https%3A//'
-        'search.creativecommons.org/photos/'
-        '29cb352c-60c1-41d8-bfa1-7d6f7d955f63'
-        '&format=xml'
-    )
-    assert response.status_code == 200
-    assert response.headers['Content-Type'] == "application/xml; charset=utf-8"
-    response_body_as_xml = ET.fromstring(response.content)
-    xml_tree = ET.ElementTree(response_body_as_xml)
-    assert xml_tree.find("width").text == '1276'
-    assert xml_tree.find("height").text == '1536'
-    assert xml_tree.find("license_url").text == 'https://creativecommons.org/licenses/by-nc-nd/4.0/'
 
 
 def test_report_endpoint():
